internal/pkg/flags: moving to internal/flags
