+++ pkg/ansible/flags/flag.go
@@ -17,7 +17,7 @@ package flags
 import (
 	"strings"
 
-	"github.com/operator-framework/operator-sdk/internal/pkg/flags/watch"
+	"github.com/operator-framework/operator-sdk/internal/flags/watch"
 	"github.com/operator-framework/operator-sdk/pkg/log/zap"
 	"github.com/spf13/pflag"
 )
